# docs/swagger.yaml
definitions:
  handlers.AccountResponse:
    properties:
      account_id:
        example: 1
        type: integer
      document_number:
        example: "12345678900"
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: account with this document number already exists
        type: string
    type: object
  handlers.ErrorResponse400:
    properties:
      error:
        example: invalid accountId provided
        type: string
    type: object
  handlers.ErrorResponse404:
    properties:
      error:
        example: account not found
        type: string
    type: object
  handlers.TransactionRequest:
    properties:
      account_id:
        example: 1
        type: integer
      amount:
        example: 123.45
        type: number
      operation_type_id:
        example: 4
        type: integer
    type: object
  handlers.TransactionResponse:
    properties:
      account_id:
        example: 1
        type: integer
      amount:
        example: 123.45
        type: number
      operation_type_id:
        example: 4
        type: integer
      transaction_id:
        example: 1
        type: integer
    type: object
info:
  contact: {}
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Creates a new account with the provided document number.
      parameters:
      - description: Document Number
        example: '"12345678900"'
        in: body
        name: document_number
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: account_id and document_number of the created account
          schema:
            $ref: '#/definitions/handlers.AccountResponse'
        "400":
          description: Invalid request or account already exists
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves the account details using the account ID.
      parameters:
      - description: Account ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: account_id and document_number of the retrieved account
          schema:
            $ref: '#/definitions/handlers.AccountResponse'
        "400":
          description: Invalid accountId provided
          schema:
            $ref: '#/definitions/handlers.ErrorResponse400'
        "404":
          description: Account not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse404'
      summary: Get an account by ID
      tags:
      - accounts
  /transactions:
    post:
      consumes:
      - application/json
      description: Creates a transaction with the provided account ID, operation type
        ID, and amount.
      parameters:
      - description: Transaction request
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/handlers.TransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: transaction_id, account_id, operation_type_id, and amount of
            the created transaction
          schema:
            $ref: '#/definitions/handlers.TransactionResponse'
        "400":
          description: Invalid request or failed transaction creation
          schema:
            $ref: '#/definitions/handlers.ErrorResponse400'
      summary: Create a new transaction
      tags:
      - transactions
swagger: "2.0"